+++ libs/hsm/tests/event_queue.c
@@ -99,7 +99,9 @@ static void hsmq_dispatch(void *ctx, const struct am_event *event) {
 static void hsmq_commit(void) {
     struct am_hsmq *me = &am_hsmq_;
     while (!am_event_queue_is_empty(&me->event_queue)) {
-        bool popped = am_event_pop_front(&me->event_queue, hsmq_dispatch, me);
+        bool popped = am_event_queue_pop_front_with_cb(
+            &me->event_queue, hsmq_dispatch, me
+        );
         AM_ASSERT(popped);
     }
 }